backend/requirements.txt
# MedScript Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.2
python-decouple==3.8
Pillow==10.0.1
numpy==1.24.4
torch==2.0.1
networkx==3.2.1
rapidfuzz==3.5.2
python-Levenshtein==0.23.0
pydantic>=2.0
